[internals] Move utils needed for tree view virtualization to shared package
