+++ packages/x-internals/src/fastMemo/index.ts
@@ -1,0 +1,1 @@
+export { fastMemo } from './fastMemo';
