+++ packages/x-internals/src/fastObjectShallowCompare/index.ts
@@ -1,0 +1,1 @@
+export { fastObjectShallowCompare } from './fastObjectShallowCompare';
