+++ packages/x-internals/src/throttle/index.ts
@@ -1,0 +1,1 @@
+export { throttle } from './throttle';
